build-target-bins: fix invalid indirect expansion
Using build-target-bins on Ubuntu 20.04 results in the following error:
"uboot_out: invalid indirect expansion"
Ubuntu 20.04 LTS uses bash version 5.0 which doesn't expand an indirect
expansion if the variable is not defined, which caused the above error.
Fix this by checking if the uboot_out variable is not empty before doing
the indirect expansion.

+++ build-target-bins.sh
@@ -321,7 +321,7 @@ do_package()
 
 				fi
 
-				if [ "${!uboot_out}" != "" ]; then
+				if [ "${uboot_out}" ] && [ "${!uboot_out}" != "" ]; then
 					# remove existing fip
 					rm -f ${PLATDIR}/${!target_name}/fip-uboot.bin
 					mkdir -p ${PLATDIR}/${!target_name}
@@ -391,7 +391,7 @@ do_package()
 				if [ "${!scp_out}" != "" ]; then
 					create_tgt_symlinks ${!scp_out} ${!target_name} "*cp_*"
 				fi
-				if [ "${!uboot_out}" != "" ]; then
+				if [ "${uboot_out}" ] && [ "${!uboot_out}" != "" ]; then
 					create_tgt_symlinks ${!uboot_out} ${!target_name} "uboot*"
 				fi
 				if [ "${!uefi_out}" != "" ]; then
